fix: extract usePollGameStatus.
